fix(model-poller): move `ModelsError` into separate file
